[03595] Refactor JobsApp.Build method into focused partial classes
- Extract hook setup methods to JobsApp.Hooks.cs
- Extract data transformation methods to JobsApp.Data.cs
- Extract DataTable configuration to JobsApp.DataTable.cs
- Extract sheet rendering to JobsApp.Sheets.cs
- Extract helper methods to JobsApp.Helpers.cs
- Reduce Build method from 441 lines to 44 lines
- Reduce cyclomatic complexity from 58 to ~10
