+++ src/Ivy.Tendril/Apps/JobsApp.cs
@@ -1,22 +1,16 @@
-using System.Reactive.Disposables;
-using System.Reactive.Linq;
-using System.Text.RegularExpressions;
-using Ivy.Tendril.Apps.Jobs;
-using Ivy.Tendril.Apps.Plans;
-using Ivy.Tendril.Models;
-using Ivy.Tendril.Helpers;
 using Ivy.Tendril.Services;
 
 namespace Ivy.Tendril.Apps;
 
 [App(title: "Jobs", icon: Icons.Activity, group: ["Apps"], order: Constants.Jobs)]
-public class JobsApp : ViewBase
+public partial class JobsApp : ViewBase
 {
     public override object Build()
     {
         var jobService = UseService<IJobService>();
         var planService = UseService<IPlanReaderService>();
         var client = UseService<IClientProvider>();
+        var config = UseService<IConfigService>();
         var refreshToken = UseRefreshToken();
         var showPlan = UseState<string?>(null);
         var showOutput = UseState<string?>(null);
@@ -27,561 +21,24 @@ public class JobsApp : ViewBase
         var streamingJobId = UseState<string?>(null);
         var hasStreamContent = UseState(false);
 
-        UseInterval(() =>
-        {
-            if (showOutput.Value is not { } activeJobId) return;
+        SetupOutputStreaming(jobService, showOutput, streamingJobId, lastProcessedIndex, hasStreamContent, outputStream);
+        SetupNotificationHooks(jobService, client);
+        SetupJobChangeHooks(jobService, refreshToken);
+        SetupAutoRefresh(jobService, refreshToken);
 
-            var activeJob = jobService.GetJob(activeJobId);
-            if (activeJob is not { Status: JobStatus.Running }) return;
-
-            var startIdx = lastProcessedIndex.Value;
-
-            if (streamingJobId.Value != activeJobId)
-            {
-                streamingJobId.Set(activeJobId);
-                hasStreamContent.Set(false);
-
-                // Immediately seed the stream with existing lines
-                var existingLines = activeJob.OutputLines.ToArray();
-                foreach (var line in existingLines)
-                {
-                    outputStream.Write(line);
-                }
-
-                if (existingLines.Length > 0 && !hasStreamContent.Value)
-                {
-                    hasStreamContent.Set(true);
-                }
-
-                lastProcessedIndex.Set(existingLines.Length);
-            }
-            else
-            {
-                // Continue pumping new lines
-                var currentLines = activeJob.OutputLines.ToArray();
-                for (var i = startIdx; i < currentLines.Length; i++)
-                {
-                    outputStream.Write(currentLines[i]);
-                }
-
-                if (currentLines.Length > 0 && !hasStreamContent.Value)
-                {
-                    hasStreamContent.Set(true);
-                }
-
-                lastProcessedIndex.Set(currentLines.Length);
-            }
-        }, TimeSpan.FromMilliseconds(100));
-        var config = UseService<IConfigService>();
-        UseEffect(() =>
-        {
-            void OnNotification(JobNotification notification)
-            {
-                if (notification.IsSuccess)
-                    client.Toast(notification.Message, notification.Title);
-                else
-                    client.Toast(notification.Message, notification.Title).Destructive();
-            }
-
-            jobService.NotificationReady += OnNotification;
-            return Disposable.Create(() => jobService.NotificationReady -= OnNotification);
-        });
-
-        UseEffect(() =>
-        {
-            void OnJobsChanged()
-            {
-                refreshToken.Refresh();
-            }
-
-            jobService.JobsStructureChanged += OnJobsChanged;
-            jobService.JobPropertyChanged += OnJobsChanged;
-            return Disposable.Create(() =>
-            {
-                jobService.JobsStructureChanged -= OnJobsChanged;
-                jobService.JobPropertyChanged -= OnJobsChanged;
-            });
-        });
-
-        UseInterval(() =>
-        {
-            var hasActiveOrRecentJobs = jobService.GetJobs().Any(j =>
-                j.Status == JobStatus.Running ||
-                (j.Status is JobStatus.Stopped or JobStatus.Failed or JobStatus.Timeout or JobStatus.Completed
-                 && j.CompletedAt.HasValue
-                 && DateTime.UtcNow - j.CompletedAt.Value < TimeSpan.FromSeconds(5)));
-
-            if (hasActiveOrRecentJobs)
-                refreshToken.Refresh();
-        }, TimeSpan.FromSeconds(5));
-
-        var updateStream = UseDataTableUpdates(
-            Observable.Interval(TimeSpan.FromSeconds(1))
-                .SelectMany(_ =>
-                {
-                    var currentJobs = jobService.GetJobs();
-                    return currentJobs
-                        .Where(j => j.Status == JobStatus.Running ||
-                                    ((j.Status is JobStatus.Stopped or JobStatus.Failed or JobStatus.Timeout or JobStatus.Completed)
-                                     && j.CompletedAt.HasValue
-                                     && DateTime.UtcNow - j.CompletedAt.Value < TimeSpan.FromMinutes(1)))
-                        .SelectMany(j => new[]
-                        {
-                            new DataTableCellUpdate(j.Id, "Timer", FormatTimer(j)),
-                            new DataTableCellUpdate(j.Id, "Cost", j.Cost.HasValue ? $"${j.Cost.Value:F2}" : ""),
-                            new DataTableCellUpdate(j.Id, "Tokens", j.Tokens.HasValue ? FormatHelper.FormatTokens(j.Tokens.Value) : ""),
-                            new DataTableCellUpdate(j.Id, "LastOutput", FormatLastOutput(j)),
-                            new DataTableCellUpdate(j.Id, "Status", j.Status),
-                            new DataTableCellUpdate(j.Id, "StatusMessage", GetStatusMessage(j))
-                        });
-                }));
-
-        var projectColors = config.Projects
-            .Select(p => new { p.Name, Color = config.GetProjectColor(p.Name) })
-            .Where(x => x.Color.HasValue)
-            .ToDictionary(x => x.Name, x => x.Color!.Value.ToString());
+        var updateStream = BuildUpdateStream(jobService);
 
         var jobs = jobService.GetJobs();
-        var rows = jobs.Select(j =>
-        {
-            var planId = ExtractPlanId(j.PlanFile);
-            if (string.IsNullOrEmpty(planId) && !string.IsNullOrEmpty(j.ReportedPlanId))
-                planId = j.ReportedPlanId;
+        var projectColors = BuildProjectColorMapping(config);
+        var rows = BuildJobRows(jobs, planService);
+        var jobsProgress = BuildStatusProgress(jobs, config);
 
-            return new JobItemRow
-            {
-                Id = j.Id,
-                Status = j.Status,
-                PlanId = planId,
-                Plan = GetPromptDisplay(j, planService),
-                Type = j.Type,
-                Project = string.Join(", ", ProjectHelper.ParseProjects(j.Project)),
-                Timer = FormatTimer(j),
-                Cost = j.Cost.HasValue ? $"${j.Cost.Value:F2}" : "",
-                Tokens = j.Tokens.HasValue ? FormatHelper.FormatTokens(j.Tokens.Value) : "",
-                LastOutput = FormatLastOutput(j),
-                LastOutputTimestamp = j.LastOutputAt,
-                StatusMessage = GetStatusMessage(j),
-                ErrorContext = j.Status is JobStatus.Failed or JobStatus.Timeout
-                    ? GetErrorContext(j)
-                    : null
-            };
-        })
-            .OrderByDescending(r => r.Id)
-            .ToList();
-
-        var statusGroups = jobs
-            .GroupBy(j => j.Status)
-            .Select(g => new { Status = g.Key, Count = g.Count() })
-            .OrderByDescending(g => g.Count)
-            .ToArray();
-
-        var statusSegments = statusGroups
-            .Select(g => new ProgressSegment(
-                g.Count,
-                GetStatusColor(g.Status),
-                g.Status.ToString()
-            ))
-            .ToArray();
-
-        var jobsProgress = new StackedProgress(statusSegments).ShowLabels();
-
-        var dataTable = rows.AsQueryable()
-            .ToDataTable(t => t.Id)
-            .RefreshToken(refreshToken)
-            .UpdateStream(updateStream)
-            .Width(Size.Full())
-            .Height(Size.Full())
-            .Header(t => t.Status, "Status")
-            .Header(t => t.Type, "Type")
-            .Header(t => t.PlanId, "Plan")
-            .Header(t => t.Plan, "Prompt/Title")
-            .Header(t => t.Project, "Project")
-            .Header(t => t.Timer, "Timer")
-            .Header(t => t.Cost, "Cost")
-            .Header(t => t.Tokens, "Tokens")
-            .Header(t => t.LastOutput, "Last Output")
-            .Header(t => t.StatusMessage, "Status")
-            .Width(t => t.Status, Size.Px(100))
-            .Width(t => t.PlanId, Size.Px(100))
-            .Width(t => t.Type, Size.Px(100))
-            .Width(t => t.Plan, Size.Px(250))
-            .Width(t => t.Project, Size.Px(100))
-            .Width(t => t.Timer, Size.Px(100))
-            .Width(t => t.LastOutput, Size.Px(100))
-            .Width(t => t.Cost, Size.Px(100))
-            .Width(t => t.Tokens, Size.Px(100))
-            .Width(t => t.StatusMessage, Size.Auto())
-            .Renderer(t => t.Status, new LabelsDisplayRenderer
-            {
-                BadgeColorMapping = StatusMappings.JobStatusColors.ToDictionary(
-                    kvp => kvp.Key.ToString(),
-                    kvp => kvp.Value.ToString()
-                )
-            })
-            .Renderer(t => t.Type, new LabelsDisplayRenderer
-            {
-                BadgeColorMapping = StatusMappings.JobTypeColors.ToDictionary(
-                    kvp => kvp.Key,
-                    kvp => kvp.Value.ToString()
-                )
-            })
-            .Renderer(t => t.Project, new LabelsDisplayRenderer
-            {
-                BadgeColorMapping = projectColors
-            })
-            .Renderer(t => t.PlanId, new LinkDisplayRenderer())
-            .Renderer(t => t.Plan, new TextDisplayRenderer())
-            .Renderer(t => t.StatusMessage, new TextDisplayRenderer())
-            .Hidden(t => t.Id)
-            .Hidden(t => t.LastOutputTimestamp)
-            .Hidden(t => t.ErrorContext)
-            .SortDirection(t => t.Id, SortDirection.Descending)
-            .Filterable(t => t.Timer, false)
-            .Filterable(t => t.LastOutput, false)
-            .Config(c =>
-            {
-                c.AllowSorting = true;
-                c.AllowFiltering = true;
-                c.ShowSearch = false;
-                c.SelectionMode = SelectionModes.None;
-                c.ShowIndexColumn = false;
-                c.BatchSize = 50;
-                c.EnableCellClickEvents = true;
-            })
-            .OnCellClick(e =>
-            {
-                if (e.Value.ColumnName == "PlanId")
-                {
-                    var planId = e.Value.CellValue?.ToString();
-                    if (!string.IsNullOrEmpty(planId))
-                    {
-                        var job = jobs.FirstOrDefault(j => ExtractPlanId(j.PlanFile) == planId);
-                        if (job != null && !string.IsNullOrEmpty(job.PlanFile))
-                        {
-                            var fullPath = Path.Combine(planService.PlansDirectory, job.PlanFile);
-                            if (Directory.Exists(fullPath))
-                                showPlan.Set(fullPath);
-                        }
-                    }
-                }
-                else if (e.Value.ColumnName == "LastOutput")
-                {
-                    var id = e.Value.RowId?.ToString();
-                    if (!string.IsNullOrEmpty(id))
-                        showOutput.Set(id);
-                }
-                else if (e.Value.ColumnName == "StatusMessage")
-                {
-                    var id = e.Value.RowId?.ToString();
-                    if (!string.IsNullOrEmpty(id))
-                    {
-                        var job = jobs.FirstOrDefault(j => j.Id == id);
-                        if (job?.Status is JobStatus.Failed or JobStatus.Timeout)
-                        {
-                            showOutput.Set(id);  // Open output sheet for failed jobs
-                        }
-                    }
-                }
-
-                return ValueTask.CompletedTask;
-            })
-            .RowActions(row =>
-            {
-                var job = jobs.FirstOrDefault(j => j.Id == row.Id);
-                var actions = new List<MenuItem>
-                {
-                    new MenuItem("View Plan", Icon: Icons.FileText, Tag: "view-plan").Tooltip("Open the associated plan"),
-                    new MenuItem("View Output", Icon: Icons.Terminal, Tag: "view-output").Tooltip(
-                        "View job output with Claude JSON rendering"),
-                    new MenuItem("Show Prompt", Icon: Icons.MessageSquare, Tag: "show-prompt").Tooltip(
-                        "Show the full prompt text"),
-                };
-
-                if (job?.Status is JobStatus.Running or JobStatus.Queued)
-                {
-                    actions.Add(new MenuItem("Stop", Icon: Icons.Square, Tag: "stop-job")
-                        .Tooltip("Stop this running job"));
-                }
-
-                if (job?.Status is JobStatus.Failed or JobStatus.Timeout or JobStatus.Stopped)
-                {
-                    actions.Add(new MenuItem("Rerun", Icon: Icons.RotateCw, Tag: "rerun-job")
-                        .Tooltip("Rerun this job"));
-                }
-
-                actions.Add(new MenuItem("Delete", Icon: Icons.Trash, Tag: "delete-job")
-                    .Tooltip("Delete this job"));
-
-                return actions.ToArray();
-            })
-            .OnRowAction(e =>
-            {
-                var tag = e.Value.Tag?.ToString();
-                var id = e.Value.Id?.ToString();
-                var job = jobs.FirstOrDefault(j => j.Id == id);
-
-                if (job != null)
-                {
-                    if (tag == "view-plan")
-                    {
-                        if (!string.IsNullOrEmpty(job.PlanFile))
-                        {
-                            var fullPath = Path.Combine(planService.PlansDirectory, job.PlanFile);
-                            if (Directory.Exists(fullPath))
-                                showPlan.Set(fullPath);
-                        }
-                    }
-                    else if (tag == "view-output")
-                    {
-                        showOutput.Set(job.Id);
-                    }
-                    else if (tag == "show-prompt")
-                    {
-                        var fullPrompt = GetFullPrompt(job);
-                        if (!string.IsNullOrEmpty(fullPrompt))
-                            showPrompt.Set(fullPrompt);
-                    }
-                    else if (tag == "stop-job")
-                    {
-                        if (job.Status is JobStatus.Running or JobStatus.Queued)
-                        {
-                            jobService.StopJob(job.Id);
-                            refreshToken.Refresh();
-                        }
-                    }
-                    else if (tag == "rerun-job")
-                    {
-                        if (job.Status is JobStatus.Failed or JobStatus.Timeout or JobStatus.Stopped)
-                        {
-                            if (job.Type == "CreatePlan" && !job.Args.Contains("-Description"))
-                            {
-                                client.Toast("Cannot rerun CreatePlan: original description was not preserved.", "Rerun Failed");
-                                return ValueTask.CompletedTask;
-                            }
-
-                            if (job.Type is "ExecutePlan" or "ExpandPlan" && job.Args.Length > 0)
-                            {
-                                var folderName = Path.GetFileName(job.Args[0]);
-                                planService.TransitionState(folderName, PlanStatus.Building);
-                            }
-                            else if (job is { Type: "UpdatePlan", Args.Length: > 0 })
-                            {
-                                var folderName = Path.GetFileName(job.Args[0]);
-                                planService.TransitionState(folderName, PlanStatus.Updating);
-                            }
-
-                            jobService.DeleteJob(job.Id);
-                            jobService.StartJob(job.Type, job.Args);
-
-                            refreshToken.Refresh();
-                        }
-                    }
-                    else if (tag == "delete-job")
-                    {
-                        if (job.Status is JobStatus.Running or JobStatus.Queued)
-                        {
-                            jobService.StopJob(job.Id);
-                        }
-
-                        jobService.DeleteJob(job.Id);
-                        refreshToken.Refresh();
-                    }
-                }
-
-                return ValueTask.CompletedTask;
-            })
-            .HeaderRight(_ => Layout.Horizontal().Gap(2)
-                              | jobsProgress
-                              | new Button().Icon(Icons.EllipsisVertical).Ghost().WithDropDown(
-                                  new MenuItem("Clear Completed", Icon: Icons.Trash, Tag: "ClearCompleted")
-                                      .OnSelect(() =>
-                                      {
-                                          jobService.ClearCompletedJobs();
-                                          refreshToken.Refresh();
-                                      }),
-                                  new MenuItem("Clear Failed", Icon: Icons.Trash, Tag: "ClearFailed").OnSelect(() =>
-                                  {
-                                      jobService.ClearFailedJobs();
-                                      refreshToken.Refresh();
-                                  })
-                              ));
+        var dataTable = BuildDataTable(rows, refreshToken, updateStream, config, planService,
+            jobService, showPlan, showOutput, showPrompt, jobs, projectColors, jobsProgress);
 
         var layout = Layout.Vertical().Height(Size.Full());
 
-        if (showPlan.Value is { } planPath)
-        {
-            var planSheetView = new PlanSheet(planPath, planService, config, openFile);
-            var fileLinkSheet = planSheetView.BuildFileLinkSheet();
-
-            var planSheet = new Sheet(
-                () => showPlan.Set(null),
-                planSheetView.Build(),
-                planSheetView.GetSheetTitle()
-            ).Width(Size.Half()).Resizable();
-
-            if (fileLinkSheet is not null) return layout | new Fragment(dataTable, planSheet, fileLinkSheet);
-
-            return layout | new Fragment(dataTable, planSheet);
-        }
-
-        if (showOutput.Value is { } jobId)
-        {
-            var outputSheetView = new OutputSheet(jobId, jobService, outputStream, hasStreamContent);
-
-            var outputSheet = new Sheet(
-                () => showOutput.Set(null),
-                outputSheetView.Build(),
-                outputSheetView.GetSheetTitle()
-            ).Width(Size.Half()).Resizable();
-
-            return layout | new Fragment(dataTable, outputSheet);
-        }
-
-        if (showPrompt.Value is { } promptText)
-        {
-            var promptSheetView = new PromptSheet(promptText);
-
-            var promptSheet = new Sheet(
-                () => showPrompt.Set(null),
-                promptSheetView.Build(),
-                "Full Prompt"
-            ).Width(Size.Half()).Resizable();
-
-            return layout | new Fragment(dataTable, promptSheet);
-        }
-
-        return layout | dataTable;
-    }
-
-    private static string? GetFullPrompt(JobItem job)
-    {
-        if (job.Type == "CreatePlan")
-        {
-            for (var i = 0; i < job.Args.Length - 1; i++)
-                if (job.Args[i].Equals("-Description", StringComparison.OrdinalIgnoreCase))
-                    return job.Args[i + 1];
-        }
-
-        return job.PlanFile;
-    }
-
-    private static string ExtractPlanId(string planFile)
-    {
-        if (string.IsNullOrEmpty(planFile)) return "";
-        var match = Regex.Match(planFile, @"^(\d{5})-");
-        return match.Success ? match.Groups[1].Value : "";
-    }
-
-    private static string FormatLastOutput(JobItem job)
-    {
-        if (job.LastOutputAt.HasValue && job.Status == JobStatus.Running)
-        {
-            var elapsed = DateTime.UtcNow - job.LastOutputAt.Value;
-            return FormatTimeSpan(elapsed);
-        }
-
-        return "-";
-    }
-
-    private static string FormatTimer(JobItem job)
-    {
-        if (job is { Status: JobStatus.Running, StartedAt: not null })
-        {
-            var elapsed = DateTime.UtcNow - job.StartedAt.Value;
-            return FormatTimeSpan(elapsed);
-        }
-
-        if (job.Status is JobStatus.Completed or JobStatus.Failed or JobStatus.Timeout or JobStatus.Stopped &&
-            job.DurationSeconds.HasValue) return FormatTimeSpan(TimeSpan.FromSeconds(job.DurationSeconds.Value));
-
-        return "-";
-    }
-
-    private static string FormatTimeSpan(TimeSpan span)
-    {
-        if (span.TotalHours >= 1)
-            return $"{(int)span.TotalHours}h {span.Minutes:D2}m";
-        return $"{span.Minutes}m {span.Seconds:D2}s";
-    }
-
-    private const int PromptDisplayMaxLength = 500;
-
-    private static string CleanPromptText(string text)
-    {
-        if (string.IsNullOrEmpty(text))
-            return text;
-        var replaced = text.Replace("\r\n", " ").Replace("\n", " ").Replace("\r", " ");
-        var collapsed = Regex.Replace(replaced, @"\s+", " ");
-        return collapsed.Trim();
-    }
-
-    private static string GetPromptDisplay(JobItem j, IPlanReaderService planService)
-    {
-        // Try to read the plan title first (for ALL job types)
-        if (!string.IsNullOrEmpty(j.PlanFile))
-        {
-            var fullPath = Path.Combine(planService.PlansDirectory, j.PlanFile);
-            var plan = planService.GetPlanByFolder(fullPath);
-            if (plan != null && !string.IsNullOrEmpty(plan.Title))
-            {
-                var title = CleanPromptText(plan.Title);
-                return title.Length > PromptDisplayMaxLength ? title[..PromptDisplayMaxLength] + "..." : title;
-            }
-        }
-
-        // Agent reported a plan title via the status API
-        if (!string.IsNullOrEmpty(j.ReportedPlanTitle))
-        {
-            var title = CleanPromptText(j.ReportedPlanTitle);
-            return title.Length > PromptDisplayMaxLength ? title[..PromptDisplayMaxLength] + "..." : title;
-        }
-
-        // CreatePlan jobs: use the -Description arg for display when no title is available yet
-        if (j.Type == "CreatePlan")
-        {
-            var desc = CleanPromptText(GetFullPrompt(j) ?? j.PlanFile);
-            return desc.Length > PromptDisplayMaxLength ? desc[..PromptDisplayMaxLength] + "..." : desc;
-        }
-
-        // Fallback to folder name
-        var pf = CleanPromptText(j.PlanFile);
-        return pf.Length > PromptDisplayMaxLength ? pf[..PromptDisplayMaxLength] + "..." : pf;
-    }
-
-    private static string GetStatusMessage(JobItem job)
-    {
-        if (!string.IsNullOrEmpty(job.StatusMessage))
-            return job.StatusMessage;
-
-        return job.Status switch
-        {
-            JobStatus.Blocked => "Waiting for dependency plan(s) to complete",
-            JobStatus.Failed => "Job encountered an error during execution",
-            JobStatus.Timeout => "Job exceeded the configured timeout",
-            JobStatus.Queued => "Waiting for a job slot to become available",
-            JobStatus.Stopped => "Job was manually stopped",
-            _ => ""
-        };
-    }
-
-    private static string? GetErrorContext(JobItem job)
-    {
-        if (job.OutputLines.Count == 0) return null;
-
-        // Return last 10 non-empty lines for detailed error view
-        var context = job.OutputLines
-            .Reverse()
-            .Where(line => !string.IsNullOrWhiteSpace(line))
-            .Take(10)
-            .Reverse()
-            .Select(JobService.SanitizeForDisplay);
-
-        return string.Join("\n", context);
-    }
-
-    private static Colors GetStatusColor(JobStatus status)
-    {
-        return StatusMappings.JobStatusColors.GetValueOrDefault(status, Colors.Slate);
+        return RenderWithSheets(dataTable, showPlan, showOutput, showPrompt, planService,
+            config, openFile, jobService, outputStream, hasStreamContent, layout);
     }
 }
